Move union query application to applyQuery, fix where clause
Fixes #9228
