+++ api/src/database/run-ast.ts
@@ -69,24 +69,7 @@ export default async function runAST(
 		);
 
 		// The actual knex query builder instance. This is a promise that resolves with the raw items from the db
-		let dbQuery = getDBQuery(schema, knex, collection, fieldNodes, query);
-
-		if (query.union) {
-			const [field, keys] = query.union;
-
-			if (keys.length) {
-				const queries = keys.map((key) => {
-					return knex.select('*').from(
-						dbQuery
-							.clone()
-							.andWhere({ [field]: key })
-							.as('foo')
-					);
-				});
-
-				dbQuery = knex.unionAll(queries);
-			}
-		}
+		const dbQuery = getDBQuery(schema, knex, collection, fieldNodes, query);
 
 		const rawItems: Item | Item[] = await dbQuery;
