+++ api/src/utils/apply-query.ts
@@ -4,7 +4,7 @@ import { customAlphabet } from 'nanoid';
 import validate from 'uuid-validate';
 import { InvalidQueryException } from '../exceptions';
 import { Relation, SchemaOverview } from '../types';
-import { Aggregate, Filter, Query } from '@directus/shared/types';
+import { Aggregate, Filter, LogicalFilterAND, Query } from '@directus/shared/types';
 import { applyFunctionToColumnName } from './apply-function-to-column-name';
 import { getColumn } from './get-column';
 import { getRelationType } from './get-relation-type';
@@ -55,7 +55,32 @@ export default function applyQuery(
 		dbQuery.offset(query.limit * (query.page - 1));
 	}
 
-	if (query.filter) {
+	if (query.union) {
+		const [field, keys] = query.union;
+
+		if (keys.length) {
+			const queries = keys.map((key) => {
+				let filter = { [field]: { _eq: key } } as Filter;
+
+				if (query.filter) {
+					if ('_and' in query.filter) {
+						(query.filter as LogicalFilterAND)._and.push(filter);
+						filter = query.filter;
+					} else {
+						filter = {
+							_and: [query.filter, filter],
+						} as LogicalFilterAND;
+					}
+				}
+
+				return knex
+					.select('*')
+					.from(applyFilter(knex, schema, dbQuery.clone(), filter, collection, subQuery).as('foo'));
+			});
+
+			dbQuery = knex.unionAll(queries);
+		}
+	} else if (query.filter) {
 		applyFilter(knex, schema, dbQuery, query.filter, collection, subQuery);
 	}
 
@@ -118,13 +143,15 @@ export function applyFilter(
 	rootFilter: Filter,
 	collection: string,
 	subQuery = false
-): void {
+) {
 	const relations: Relation[] = schema.relations;
 
 	const aliasMap: Record<string, string> = {};
 
 	addJoins(rootQuery, rootFilter, collection);
 	addWhereClauses(knex, rootQuery, rootFilter, collection);
+
+	return rootQuery;
 
 	function addJoins(dbQuery: Knex.QueryBuilder, filter: Filter, collection: string) {
 		for (const [key, value] of Object.entries(filter)) {
